Move `ReferenceFrame` to feature types header
