+++ common/include/pcl/point_types.h
@@ -184,11 +184,6 @@ namespace pcl
     */
   struct PointWithViewpoint;
 
-  /** \brief Members: Axis x_axis, y_axis, z_axis
-    * \ingroup common
-    */
-  struct ReferenceFrame;
-
   /** \brief Members: float descriptor[1980], rf[9]
     * \ingroup common
     */
