+++ common/src/feature_types.cpp
@@ -56,4 +56,14 @@ namespace pcl
     os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 1343 ? ", " : ")");
     return (os);
   }
+
+  std::ostream&
+  operator << (std::ostream& os, const ReferenceFrame& p)
+  {
+    os << "("
+       << p.x_axis[0] << " " << p.x_axis[1] << " " << p.x_axis[2] << ","
+       << p.y_axis[0] << " " << p.y_axis[1] << " " << p.y_axis[2] << ","
+       << p.z_axis[0] << " " << p.z_axis[1] << " " << p.z_axis[2] << ")";
+    return (os);
+  }
 } // namespace pcl
